move article_class.py to article.py; update imports
